+++ lib/vcloud/edge_gateway/edge_gateway_configuration.rb
@@ -2,10 +2,10 @@ module Vcloud
   module EdgeGateway
     class EdgeGatewayConfiguration
 
-      attr_reader :config
+      attr_reader :config, :diff
 
       def initialize(local_config, remote_config, edge_gateway_interfaces)
-        @config = generate_new_config(local_config, remote_config, edge_gateway_interfaces)
+        @config, @diff = generate_new_config(local_config, remote_config, edge_gateway_interfaces)
       end
 
       def update_required?
@@ -15,15 +15,19 @@ module Vcloud
       private
       def generate_new_config(local_config, remote_config, edge_gateway_interfaces)
         new_config = { }
+        diff = { }
 
         firewall_service_config =
           EdgeGateway::ConfigurationGenerator::FirewallService.new.
             generate_fog_config(local_config[:firewall_service])
 
         unless firewall_service_config.nil?
-          differ = EdgeGateway::FirewallConfigurationDiffer.
-            new(firewall_service_config, remote_config[:FirewallService])
+          differ = EdgeGateway::FirewallConfigurationDiffer.new(
+            remote_config[:FirewallService],
+            firewall_service_config
+          )
           unless differ.diff.empty?
+            diff[:FirewallService] = differ.diff
             new_config[:FirewallService] = firewall_service_config
           end
         end
@@ -34,8 +38,12 @@ module Vcloud
         ).generate_fog_config
 
         unless nat_service_config.nil?
-          differ = EdgeGateway::NatConfigurationDiffer.new(nat_service_config, remote_config[:NatService])
+          differ = EdgeGateway::NatConfigurationDiffer.new(
+            remote_config[:NatService],
+            nat_service_config
+          )
           unless differ.diff.empty?
+            diff[:NatService] = differ.diff
             new_config[:NatService] = nat_service_config
           end
         end
@@ -47,15 +55,16 @@ module Vcloud
 
         unless load_balancer_service_config.nil?
           differ = EdgeGateway::LoadBalancerConfigurationDiffer.new(
-                     load_balancer_service_config,
-                     remote_config[:LoadBalancerService]
-                   )
+            remote_config[:LoadBalancerService],
+            load_balancer_service_config
+          )
           unless differ.diff.empty?
+            diff[:LoadBalancerService] = differ.diff
             new_config[:LoadBalancerService] = load_balancer_service_config
           end
         end
 
-        new_config
+        return new_config, diff
       end
 
     end